Refactor templateArea into its own component, add some animation to smooth out the movement.
